# api-spec.yaml
openapi: 3.0.3
info:
  title: Crealia SaaS API
  description: API complète pour la plateforme SaaS de création et d'analyse de contenu
  version: 1.0.0
  contact:
    name: Crealia Support
    email: support@crealia.com
servers:
  - url: https://api.crealia.com/v1
    description: Production server
  - url: https://staging-api.crealia.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /ai/generate:
    post:
      tags: [AI Art]
      summary: Generate AI art
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIGenerationRequest'
      responses:
        '202':
          description: AI generation job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    AIGenerationRequest:
      type: object
      required: [type, prompt]
      properties:
        type:
          $ref: '#/components/schemas/AIArtType'
        prompt:
          type: string
        negativePrompt:
          type: string
        model:
          type: string
          default: "stable-diffusion-xl"
        parameters:
          type: object
          properties:
            seed:
              type: integer
            steps:
              type: integer
              default: 20
            cfgScale:
              type: number
              default: 7.5
            width:
              type: integer
              default: 1024
            height:
              type: integer
              default: 1024
        projectId:
          type: string
    AIArtType:
      type: string
      enum: [TEXT_TO_IMAGE, IMAGE_TO_IMAGE, INPAINTING, OUTPAINTING, STYLE_TRANSFER, UPSCALE, ENHANCE, AVATAR_GENERATION, BATCH_GENERATION]

    # Video Editor
...
    Job:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/JobType'
        status:
          $ref: '#/components/schemas/JobStatus'
        priority:
          $ref: '#/components/schemas/JobPriority'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        stage:
          type: string
        input:
          type: object
        output:
          type: object
        error:
          type: string
        retryCount:
          type: integer
        maxRetries:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        duration:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobType:
      type: string
      enum: [ASSET_PROCESSING, THUMBNAIL_GENERATION, EDIT_APPLICATION, EXPORT_GENERATION, AI_INFERENCE, BACKGROUND_REMOVAL, FACE_DETECTION, STYLE_TRANSFER, UPSCALE, VIDEO_TRANSCODE, AUDIO_PROCESSING, METADATA_EXTRACTION, BATCH_PROCESSING]
    JobStatus:
      type: string
      enum: [PENDING, QUEUED, PROCESSING, COMPLETED, FAILED, CANCELLED, STALLED, DELAYED]
    JobPriority:
      type: string
      enum: [LOW, NORMAL, HIGH, URGENT]

    # User
...